Refactor

+++ src/LNSF.Application/Services/AccountService.cs
@@ -11,29 +11,21 @@ public class AccountService
 {
     private readonly IAccountRepository _accountRepository;
     private readonly AccountValidator _accountValidator;
-    private readonly AccountFilterValidator _accountFilterValidator;
     private readonly PasswordValidator _passwordValidator;
 
     public AccountService(IAccountRepository accountRepository,
         AccountValidator accountValidator,
-        AccountFilterValidator accountFilterValidator,
         PasswordValidator passwordValidator)
     {
         _accountRepository = accountRepository;
         _accountValidator = accountValidator;
-        _accountFilterValidator = accountFilterValidator;
         _passwordValidator = passwordValidator;
     }
 
     public async Task<List<Account>> Query(AccountFilter filter)
-    {
-        var validationResult = await _accountFilterValidator.ValidateAsync(filter);
-
-        if (!validationResult.IsValid) throw new AppException(validationResult.ToString(), HttpStatusCode.BadRequest);
-        
+    {        
         var accounts = await _accountRepository.Query(filter);
         accounts.ForEach(x => x.Password = "");
-
         return accounts;
     }
 
@@ -41,55 +33,52 @@ public class AccountService
     {
         var account = await _accountRepository.Get(userName, password);
         account.Password = "";
-
         return account;
     }
 
     public async Task<Account> Create(Account account)
     {
         var validationResult = await _accountValidator.ValidateAsync(account);
-
-        if (!validationResult.IsValid) throw new AppException(validationResult.ToString(), HttpStatusCode.BadRequest);
+        if (!validationResult.IsValid) 
+            throw new AppException(validationResult.ToString(), HttpStatusCode.BadRequest);
 
         account.Id = 0;
         account = await _accountRepository.Post(account);
         account.Password = "";
-
-        return account;
-    }
-
-    public async Task<Account> Update(Account account)
-    {
-        var validationResult = await _accountValidator.ValidateAsync(account);
-
-        if (!validationResult.IsValid) throw new AppException(validationResult.ToString(), HttpStatusCode.BadRequest);
-
-        await _accountRepository.Get(account.Id);
-
-        account = await _accountRepository.Put(account);
-        account.Password = "";
-
         return account;
     }
 
     public async Task<Account> Update(Account account, string oldPassword)
     {
         var validationResult = await _passwordValidator.ValidateAsync(oldPassword);
-
-        if (!validationResult.IsValid) throw new AppException(validationResult.ToString(), HttpStatusCode.BadRequest);
+        if (!validationResult.IsValid) 
+            throw new AppException(validationResult.ToString(), HttpStatusCode.BadRequest);
 
         account.Password = oldPassword;
-
         return await Update(account);
+    }
+
+    public async Task<Account> Update(Account account)
+    {
+        var validationResult = await _accountValidator.ValidateAsync(account);
+        if (!validationResult.IsValid) 
+            throw new AppException(validationResult.ToString(), HttpStatusCode.BadRequest);
+        if (!await _accountRepository.Exists(account.Id)) 
+            throw new AppException("Account not Found!", HttpStatusCode.NotFound); 
+
+        account = await _accountRepository.Put(account);
+        account.Password = "";
+        return account;
     }
 
     public async Task<Account> Delete(int id)
     {
-        var account = await _accountRepository.Get(id);
+        if (!await _accountRepository.Exists(id)) 
+            throw new AppException("AccountId not Found!", HttpStatusCode.NotFound);
 
+        var account = await _accountRepository.Get(id);
         account = await _accountRepository.Delete(account);
         account.Password = "";
-
         return account;
     }
 }
